refactor: unify user and user_preferences modules
