--- components/renku_data_services/user_preferences/__init__.py
@@ -1,1 +1,0 @@
-"""Blueprints for user preferences."""
