--- components/renku_data_services/user_preferences/api.spec.yaml
@@ -1,153 +1,0 @@
----
-openapi: 3.0.2
-info:
-  title: Renku Data Services API
-  description: |
-    This service is the main backend for Renku. It provides information about users, projects,
-    cloud storage, access to compute resources and many other things.
-  version: v1
-servers:
-  - url: /api/data
-  - url: /ui-server/api/data
-paths:
-  "/user/preferences":
-    get:
-      summary: Get user preferences for the currently logged in user
-      responses:
-        "200":
-          description: The user preferences
-          content:
-            "application/json":
-              schema:
-                $ref: "#/components/schemas/UserPreferences"
-        "404":
-          description: The user has no preferences saved
-          content:
-            "application/json":
-              schema:
-                $ref: "#/components/schemas/ErrorResponse"
-        default:
-          $ref: "#/components/responses/Error"
-      tags:
-        - user_preferences
-  "/user/preferences/pinned_projects":
-    post:
-      summary: Add a pinned project
-      requestBody:
-        required: true
-        content:
-          "application/json":
-            schema:
-              $ref: "#/components/schemas/AddPinnedProject"
-      responses:
-        "200":
-          description: The updated user preferences
-          content:
-            "application/json":
-              schema:
-                $ref: "#/components/schemas/UserPreferences"
-        default:
-          $ref: "#/components/responses/Error"
-      tags:
-        - user_preferences
-    delete:
-      summary: Remove one or all pinned projects
-      parameters:
-        - in: query
-          description: query parameters
-          name: delete_pinned_params
-          style: form
-          explode: true
-          schema:
-            type: object
-            additionalProperties: false
-            properties:
-              project_slug:
-                type: string
-                default: ""
-      responses:
-        "200":
-          description: The updated user preferences
-          content:
-            "application/json":
-              schema:
-                $ref: "#/components/schemas/UserPreferences"
-        default:
-          $ref: "#/components/responses/Error"
-      tags:
-        - user_preferences
-
-components:
-  schemas:
-    UserPreferences:
-      type: object
-      description: The object containing user preferences
-      additionalProperties: false
-      properties:
-        user_id:
-          $ref: "#/components/schemas/UserId"
-        pinned_projects:
-          $ref: "#/components/schemas/PinnedProjects"
-      required: ["user_id", "pinned_projects"]
-    UserId:
-      type: string
-      description: Keycloak user ID
-      example: f74a228b-1790-4276-af5f-25c2424e9b0c
-      pattern: "^[A-Za-z0-9]{1}[A-Za-z0-9-]+$"
-    PinnedProjects:
-      type: object
-      description: The list of projects a user has pinned on their dashboard
-      properties:
-        project_slugs:
-          type: array
-          items:
-            $ref: "#/components/schemas/ProjectSlug"
-    ProjectSlug:
-      type: string
-      description: The slug used to identify a project
-      minLength: 3
-      example: "user/my-project"
-      # limitations based on allowed characters in project slugs from Gitlab from here:
-      # https://docs.gitlab.com/ee/user/reserved_names.html
-      pattern: "[a-zA-Z0-9_.-/]"
-    AddPinnedProject:
-      type: object
-      additionalProperties: false
-      properties:
-        project_slug:
-          $ref: "#/components/schemas/ProjectSlug"
-      required: ["project_slug"]
-    ErrorResponse:
-      type: object
-      properties:
-        error:
-          type: object
-          properties:
-            code:
-              type: integer
-              minimum: 0
-              exclusiveMinimum: true
-              example: 1404
-            detail:
-              type: string
-              example: "A more detailed optional message showing what the problem was"
-            message:
-              type: string
-              example: "Something went wrong - please try again later"
-          required: ["code", "message"]
-      required: ["error"]
-
-  responses:
-    Error:
-      description: The schema for all 4xx and 5xx responses
-      content:
-        "application/json":
-          schema:
-            $ref: "#/components/schemas/ErrorResponse"
-  securitySchemes:
-    oidc:
-      type: openIdConnect
-      openIdConnectUrl: /auth/realms/Renku/.well-known/openid-configuration
-security:
-  - oidc:
-      - openid
