--- components/renku_data_services/user_preferences/apispec_base.py
@@ -1,12 +1,0 @@
-"""Base models for API specifications."""
-
-from pydantic import BaseModel
-
-
-class BaseAPISpec(BaseModel):
-    """Base API specification."""
-
-    class Config:
-        """Enables orm mode for pydantic."""
-
-        from_attributes = True
