--- components/renku_data_services/user_preferences/blueprints.py
@@ -1,57 +1,0 @@
-"""User preferences app."""
-
-from dataclasses import dataclass
-
-from sanic import Request, json
-from sanic.response import JSONResponse
-from sanic_ext import validate
-
-import renku_data_services.base_models as base_models
-from renku_data_services.base_api.auth import authenticate
-from renku_data_services.base_api.blueprint import BlueprintFactoryResponse, CustomBlueprint
-from renku_data_services.base_api.misc import validate_query
-from renku_data_services.user_preferences import apispec, models
-from renku_data_services.user_preferences.db import UserPreferencesRepository
-
-
-@dataclass(kw_only=True)
-class UserPreferencesBP(CustomBlueprint):
-    """Handlers for manipulating user preferences."""
-
-    user_preferences_repo: UserPreferencesRepository
-    authenticator: base_models.Authenticator
-
-    def get(self) -> BlueprintFactoryResponse:
-        """Get user preferences for the logged in user."""
-
-        @authenticate(self.authenticator)
-        async def _get(_: Request, user: base_models.APIUser) -> JSONResponse:
-            user_preferences: models.UserPreferences | None
-            user_preferences = await self.user_preferences_repo.get_user_preferences(user=user)
-            return json(apispec.UserPreferences.model_validate(user_preferences).model_dump())
-
-        return "/user/preferences", ["GET"], _get
-
-    def post_pinned_projects(self) -> BlueprintFactoryResponse:
-        """Add a pinned project to user preferences for the logged in user."""
-
-        @authenticate(self.authenticator)
-        @validate(json=apispec.AddPinnedProject)
-        async def _post(_: Request, user: base_models.APIUser, body: apispec.AddPinnedProject) -> JSONResponse:
-            res = await self.user_preferences_repo.add_pinned_project(user=user, project_slug=body.project_slug)
-            return json(apispec.UserPreferences.model_validate(res).model_dump())
-
-        return "/user/preferences/pinned_projects", ["POST"], _post
-
-    def delete_pinned_projects(self) -> BlueprintFactoryResponse:
-        """Remove a pinned project from user preferences for the logged in user."""
-
-        @authenticate(self.authenticator)
-        @validate_query(query=apispec.DeletePinnedParams)
-        async def _delete(
-            request: Request, user: base_models.APIUser, query: apispec.DeletePinnedParams
-        ) -> JSONResponse:
-            res = await self.user_preferences_repo.remove_pinned_project(user=user, project_slug=query.project_slug)
-            return json(apispec.UserPreferences.model_validate(res).model_dump())
-
-        return "/user/preferences/pinned_projects", ["DELETE"], _delete
