+++ components/renku_data_services/users/apispec.py
@@ -1,6 +1,6 @@
 # generated by datamodel-codegen:
 #   filename:  api.spec.yaml
-#   timestamp: 2024-08-06T05:55:30+00:00
+#   timestamp: 2024-08-06T06:08:47+00:00
 
 from __future__ import annotations
 
@@ -26,6 +26,29 @@ class Version(BaseAPISpec):
 class SecretKind(Enum):
     general = "general"
     storage = "storage"
+
+
+class ProjectSlug(RootModel[str]):
+    root: str = Field(
+        ...,
+        description="The slug used to identify a project",
+        example="user/my-project",
+        min_length=3,
+        pattern="[a-zA-Z0-9_.-/]",
+    )
+
+
+class AddPinnedProject(BaseAPISpec):
+    model_config = ConfigDict(
+        extra="forbid",
+    )
+    project_slug: str = Field(
+        ...,
+        description="The slug used to identify a project",
+        example="user/my-project",
+        min_length=3,
+        pattern="[a-zA-Z0-9_.-/]",
+    )
 
 
 class Error(BaseAPISpec):
@@ -64,6 +87,17 @@ class UserSecretsParams(BaseAPISpec):
 
 class UserSecretsGetParametersQuery(BaseAPISpec):
     user_secrets_params: Optional[UserSecretsParams] = None
+
+
+class DeletePinnedParams(BaseAPISpec):
+    model_config = ConfigDict(
+        extra="forbid",
+    )
+    project_slug: str = ""
+
+
+class UserPreferencesPinnedProjectsDeleteParametersQuery(BaseAPISpec):
+    delete_pinned_params: Optional[DeletePinnedParams] = None
 
 
 class UserWithId(BaseAPISpec):
@@ -166,5 +200,22 @@ class SecretPatch(BaseAPISpec):
     )
 
 
+class PinnedProjects(BaseAPISpec):
+    project_slugs: Optional[List[ProjectSlug]] = None
+
+
 class SecretsList(RootModel[List[SecretWithId]]):
     root: List[SecretWithId] = Field(..., description="A list of secrets", min_length=0)
+
+
+class UserPreferences(BaseAPISpec):
+    model_config = ConfigDict(
+        extra="forbid",
+    )
+    user_id: str = Field(
+        ...,
+        description="Keycloak user ID",
+        example="f74a228b-1790-4276-af5f-25c2424e9b0c",
+        pattern="^[A-Za-z0-9]{1}[A-Za-z0-9-]+$",
+    )
+    pinned_projects: PinnedProjects
