+++ projects/background_jobs/pyproject.toml
@@ -18,7 +18,6 @@ packages = [
   { include = "renku_data_services/message_queue", from = "../../components" },
   { include = "renku_data_services/db_config", from = "../../components" },
   { include = "renku_data_services/k8s", from = "../../components" },
-  { include = "renku_data_services/user_preferences", from = "../../components" },
   { include = "renku_data_services/crc", from = "../../components" },
   { include = "renku_data_services/project", from = "../../components" },
   { include = "renku_data_services/authz", from = "../../components" },
