+++ pyproject.toml
@@ -19,7 +19,6 @@ packages = [
   { include = "renku_data_services/base_models", from = "components" },
   { include = "renku_data_services/base_api", from = "components" },
   { include = "renku_data_services/storage", from = "components" },
-  { include = "renku_data_services/user_preferences", from = "components" },
   { include = "renku_data_services/utils", from = "components" },
   { include = "renku_data_services/git", from = "components" },
   { include = "renku_data_services/users", from = "components" },
@@ -195,7 +194,6 @@ module = [
   "renku_data_services.repositories.apispec",
   "renku_data_services.secrets.apispec",
   "renku_data_services.session.apispec",
-  "renku_data_services.user_preferences.apispec",
   "renku_data_services.users.apispec",
   "renku_data_services.data_api.error_handler",
   "renku_data_services.namespace.apispec",
